Read GT3X File

Read GT3X File

py_read_gt3x(path, create_time = FALSE, verbose = FALSE)

Arguments

path

file to GT3X file

create_time

Should time stamps be created?

verbose

print diagnostic messages

Value

A list of meta data, accelerometry data, and time stamps

Examples

reticulate::py_config()
#> python:         /usr/local/bin/python3
#> libpython:      /usr/local/opt/python@3.9/Frameworks/Python.framework/Versions/3.9/lib/python3.9/config-3.9-darwin/libpython3.9.dylib
#> pythonhome:     /usr/local/Cellar/python@3.9/3.9.10/Frameworks/Python.framework/Versions/3.9:/usr/local/Cellar/python@3.9/3.9.10/Frameworks/Python.framework/Versions/3.9
#> version:        3.9.10 (main, Jan 15 2022, 11:48:00)  [Clang 13.0.0 (clang-1300.0.29.3)]
#> numpy:           [NOT FOUND]
#> 
#> python versions found: 
#>  /usr/local/bin/python3
#>  /usr/local/miniconda/bin/python

path = system.file("extdata", "TAS1H30182785_2019-09-17.gt3x",
package = "pygt3x")
if (pygt3x:::have_python_requirements()) {
res = py_read_gt3x(path)
}
#> Warning: numpy, bitstring packages not found, please try to install using reticulate::py_install(c("numpy", "bitstring"))
#> pygt3x may not work

if (FALSE) {
if (pygt3x:::have_python_requirements()) {
  out = impute_zeros(res$data, res$dates, res$header)
  out = impute_zeros(res)

  url = "https://github.com/THLfi/read.gt3x/files/3522749/GT3X%2B.01.day.gt3x.zip"
  destfile = tempfile(fileext = ".zip")
  dl = download.file(url, destfile = destfile, mode = "wb")
  gt3x_file = unzip(destfile, exdir = tempdir())
  gt3x_file = gt3x_file[!grepl("__MACOSX", gt3x_file)]
  path = gt3x_file

  res = py_read_gt3x(path)
  df = impute_zeros(res$data, res$dates, res$header)
}
}